Décorrele la dépendance circulaire entre scores et snapshots
