+++ backend/src/personnalisations/models/get-personnalisation-reponses.response.ts
@@ -1,8 +1,9 @@
 import { z } from 'zod';
 
-export const getPersonnalitionReponsesResponseSchema = z
+export const personnalisationReponsesPayloadSchema = z
   .record(z.string(), z.union([z.string(), z.number(), z.boolean(), z.null()]))
   .describe('Reponses aux questions de personnalisation');
-export type GetPersonnalisationReponsesResponseType = z.infer<
-  typeof getPersonnalitionReponsesResponseSchema
+
+export type PersonnalisationReponsesPayload = z.infer<
+  typeof personnalisationReponsesPayloadSchema
 >;
