+++ backend/src/referentiels/compute-score/compute-score-mode.enum.ts
