+++ backend/src/referentiels/compute-score/scores.controller.ts
@@ -1,22 +1,11 @@
-import { COLLECTIVITE_ID_PARAM_KEY } from '@/backend/collectivites/shared/models/collectivite-api.constants';
 import { createZodDto } from '@anatine/zod-nestjs';
-import {
-  Controller,
-  Delete,
-  Get,
-  Logger,
-  Next,
-  Param,
-  Query,
-  Req,
-  Res,
-} from '@nestjs/common';
+import { Controller, Get, Logger, Param, Query, Req } from '@nestjs/common';
 import { ApiExcludeEndpoint, ApiResponse, ApiTags } from '@nestjs/swagger';
-import { NextFunction, Request, Response } from 'express';
+import { Request } from 'express';
+import z from 'zod';
 import { AllowAnonymousAccess } from '../../auth/decorators/allow-anonymous-access.decorator';
 import { TokenInfo } from '../../auth/decorators/token-info.decorators';
-import { AuthenticatedUser, AuthUser } from '../../auth/models/auth.models';
-import ExportScoreService from '../export-score/export-score.service';
+import { AuthenticatedUser } from '../../auth/models/auth.models';
 import { checkMultipleReferentielScoresRequestSchema } from '../models/check-multiple-referentiel-scores.request';
 import { CheckReferentielScoresRequestType } from '../models/check-referentiel-scores.request';
 import { getActionStatutsRequestSchema } from '../models/get-action-statuts.request';
@@ -24,30 +13,21 @@ import { GetCheckScoresResponseType } from '../models/get-check-scores.response'
 import { GetMultipleCheckScoresResponseType } from '../models/get-multiple-check-scores.response';
 import { getReferentielMultipleScoresRequestSchema } from '../models/get-referentiel-multiple-scores.request';
 import { getReferentielMultipleScoresResponseSchema } from '../models/get-referentiel-multiple-scores.response';
-import { getReferentielScoresRequestSchema } from '../models/get-referentiel-scores.request';
-import { getScoreSnapshotRequestSchema } from '../models/get-score-snapshot.request';
-import { getScoreSnapshotsRequestSchema } from '../models/get-score-snapshots.request';
-import { getScoreSnapshotsResponseSchema } from '../models/get-score-snapshots.response';
 import {
-  REFERENTIEL_ID_PARAM_KEY,
-  SNAPSHOT_REF_PARAM_KEY,
-} from '../models/referentiel-api.constants';
-import { ReferentielId } from '../models/referentiel-id.enum';
-import { ListSnapshotsService } from '../snapshots/list-snapshots/list-snapshots.service';
-import { SnapshotsService } from '../snapshots/snapshots.service';
+  ReferentielId,
+  referentielIdEnumSchema,
+} from '../models/referentiel-id.enum';
+import { snapshotJalonEnumSchema } from '../snapshots/snapshot-jalon.enum';
+import { snapshotWithoutPayloadsSchema } from '../snapshots/snapshot.table';
 import { actionStatutsByActionIdSchema } from './action-statuts-by-action-id.dto';
-import {
-  getReferentielScoresResponseSchema,
-  GetReferentielScoresResponseType,
-} from './get-referentiel-scores.response';
 import ScoresService from './scores.service';
 
-class GetReferentielScoresRequestClass extends createZodDto(
-  getReferentielScoresRequestSchema
-) {}
-class GetReferentielScoresResponseClass extends createZodDto(
-  getReferentielScoresResponseSchema
-) {}
+// class GetReferentielScoresRequestClass extends createZodDto(
+//   getReferentielScoresRequestSchema
+// ) {}
+// class GetReferentielScoresResponseClass extends createZodDto(
+//   scoresPayloadSchema
+// ) {}
 
 class GetReferentielMultipleScoresRequestClass extends createZodDto(
   getReferentielMultipleScoresRequestSchema
@@ -69,29 +49,38 @@ class CheckMultipleReferentielScoresRequestClass extends createZodDto(
   checkMultipleReferentielScoresRequestSchema
 ) {}
 
-class GetScoreSnapshotsRequestClass extends createZodDto(
-  getScoreSnapshotsRequestSchema
-) {}
+// class GetScoreSnapshotsRequestClass extends createZodDto(
+//   getScoreSnapshotsRequestSchema
+// ) {}
 
-class GetScoreSnapshotsResponseClass extends createZodDto(
-  getScoreSnapshotsResponseSchema
-) {}
+const getScoreSnapshotsResponseSchema = z
+  .object({
+    collectiviteId: z.number(),
+    referentielId: referentielIdEnumSchema,
+    typesJalon: snapshotJalonEnumSchema.array(),
+    snapshots: snapshotWithoutPayloadsSchema
+      .omit({
+        referentielId: true,
+        collectiviteId: true,
+      })
+      .array(),
+  })
+  .describe(
+    'Liste des snapshots de score pour une collectivité et un référentiel'
+  );
 
-class GetScoreSnapshotRequestClass extends createZodDto(
-  getScoreSnapshotRequestSchema
-) {}
+export type GetScoreSnapshotsResponseType = z.infer<
+  typeof getScoreSnapshotsResponseSchema
+>;
+
+// class GetScoreSnapshotsResponseClass extends createZodDto(
+//   getScoreSnapshotsResponseSchema
+// ) {}
 
 @ApiTags('Referentiels')
 @Controller('')
 export class ReferentielsScoringController {
-  private readonly logger = new Logger(ReferentielsScoringController.name);
-
-  constructor(
-    private readonly referentielsScoringService: ScoresService,
-    private readonly referentielsScoringSnapshotsService: SnapshotsService,
-    private readonly listSnapshots: ListSnapshotsService,
-    private readonly exportReferentielScoreService: ExportScoreService
-  ) {}
+  constructor(private readonly scoresService: ScoresService) {}
 
   @AllowAnonymousAccess()
   @Get(
@@ -104,7 +93,7 @@ export class ReferentielsScoringController {
     @Query() parameters: GetActionStatutsRequestClass,
     @TokenInfo() tokenInfo: AuthenticatedUser
   ): Promise<GetActionStatutsResponseClass> {
-    return this.referentielsScoringService.getReferentielActionStatuts(
+    return this.scoresService.getReferentielActionStatuts(
       referentielId,
       collectiviteId,
       parameters.date,
@@ -112,37 +101,41 @@ export class ReferentielsScoringController {
     );
   }
 
-  @AllowAnonymousAccess()
-  @Get('referentiels/:referentiel_id/scores')
-  @ApiResponse({ type: GetReferentielMultipleScoresResponseClass })
-  async getReferentielMultipleScorings(
-    @Param('referentiel_id') referentielId: ReferentielId,
-    @Query() parameters: GetReferentielMultipleScoresRequestClass,
-    @TokenInfo() tokenInfo: AuthenticatedUser
-  ): Promise<GetReferentielMultipleScoresResponseClass> {
-    return await this.referentielsScoringService.computeScoreForMultipleCollectivite(
-      referentielId,
-      parameters,
-      tokenInfo
-    );
-  }
+  // TODO: À priori non utilisé. À supprimer après confirmation en prod.
+  //
+  // @AllowAnonymousAccess()
+  // @Get('referentiels/:referentiel_id/scores')
+  // @ApiResponse({ type: GetReferentielMultipleScoresResponseClass })
+  // async getReferentielMultipleScorings(
+  //   @Param('referentiel_id') referentielId: ReferentielId,
+  //   @Query() parameters: GetReferentielMultipleScoresRequestClass,
+  //   @TokenInfo() tokenInfo: AuthenticatedUser
+  // ): Promise<GetReferentielMultipleScoresResponseClass> {
+  //   return await this.scoresService.computeScoreForMultipleCollectivite(
+  //     referentielId,
+  //     parameters,
+  //     tokenInfo
+  //   );
+  // }
 
-  @AllowAnonymousAccess()
-  @Get('collectivites/:collectivite_id/referentiels/:referentiel_id/scores')
-  @ApiResponse({ type: GetReferentielScoresResponseClass })
-  async getReferentielScoring(
-    @Param('collectivite_id') collectiviteId: number,
-    @Param('referentiel_id') referentielId: ReferentielId,
-    @Query() parameters: GetReferentielScoresRequestClass,
-    @TokenInfo() tokenInfo: AuthenticatedUser
-  ): Promise<GetReferentielScoresResponseClass> {
-    return await this.referentielsScoringService.computeScoreForCollectivite(
-      referentielId,
-      collectiviteId,
-      parameters,
-      tokenInfo
-    );
-  }
+  // TODO: À priori non utilisé. À supprimer après confirmation en prod.
+  //
+  // @AllowAnonymousAccess()
+  // @Get('collectivites/:collectivite_id/referentiels/:referentiel_id/scores')
+  // @ApiResponse({ type: GetReferentielScoresResponseClass })
+  // async getReferentielScoring(
+  //   @Param('collectivite_id') collectiviteId: number,
+  //   @Param('referentiel_id') referentielId: ReferentielId,
+  //   @Query() parameters: GetReferentielScoresRequestClass,
+  //   @TokenInfo() tokenInfo: AuthenticatedUser
+  // ): Promise<GetReferentielScoresResponseClass> {
+  //   return await this.scoresService.computeScoreForCollectivite(
+  //     referentielId,
+  //     collectiviteId,
+  //     parameters,
+  //     tokenInfo
+  //   );
+  // }
 
   @AllowAnonymousAccess()
   @ApiExcludeEndpoint() // Not in documentation
@@ -156,7 +149,7 @@ export class ReferentielsScoringController {
     @Req() req: Request
   ): Promise<GetCheckScoresResponseType> {
     const checkReferentielHostUrl = `${req.protocol}://${req.get('Host')}`;
-    return await this.referentielsScoringService.checkScoreForCollectivite(
+    return await this.scoresService.checkScoreForCollectivite(
       referentielId,
       collectiviteId,
       parameters,
@@ -164,93 +157,73 @@ export class ReferentielsScoringController {
     );
   }
 
-  @AllowAnonymousAccess()
-  @Get(
-    `collectivites/:${COLLECTIVITE_ID_PARAM_KEY}/referentiels/:${REFERENTIEL_ID_PARAM_KEY}/score-snapshots`
-  )
-  @ApiResponse({ type: GetScoreSnapshotsResponseClass })
-  async list(
-    @Param(COLLECTIVITE_ID_PARAM_KEY) collectiviteId: number,
-    @Param(REFERENTIEL_ID_PARAM_KEY) referentielId: ReferentielId,
-    @Query() options: GetScoreSnapshotsRequestClass
-  ): Promise<GetScoreSnapshotsResponseClass> {
-    return this.listSnapshots.list({
-      collectiviteId,
-      referentielId,
-      options: {
-        jalons: options.typesJalon,
-      },
-    });
-  }
+  // TODO: À priori non utilisé. À supprimer après confirmation en prod.
+  //
+  // @AllowAnonymousAccess()
+  // @Get(
+  //   `collectivites/:${COLLECTIVITE_ID_PARAM_KEY}/referentiels/:${REFERENTIEL_ID_PARAM_KEY}/score-snapshots`
+  // )
+  // @ApiResponse({ type: GetScoreSnapshotsResponseClass })
+  // async list(
+  //   @Param(COLLECTIVITE_ID_PARAM_KEY) collectiviteId: number,
+  //   @Param(REFERENTIEL_ID_PARAM_KEY) referentielId: ReferentielId,
+  //   @Query() options: GetScoreSnapshotsRequestClass
+  // ): Promise<GetScoreSnapshotsResponseClass> {
+  //   return this.listSnapshots.list({
+  //     collectiviteId,
+  //     referentielId,
+  //     options: {
+  //       jalons: options.typesJalon,
+  //     },
+  //   });
+  // }
 
-  @AllowAnonymousAccess()
-  @Get(
-    `collectivites/:${COLLECTIVITE_ID_PARAM_KEY}/referentiels/:${REFERENTIEL_ID_PARAM_KEY}/score-snapshots/:${SNAPSHOT_REF_PARAM_KEY}`
-  )
-  async getReferentielScoreSnapshot(
-    @Param(COLLECTIVITE_ID_PARAM_KEY) collectiviteId: number,
-    @Param(REFERENTIEL_ID_PARAM_KEY) referentielId: ReferentielId,
-    @Param(SNAPSHOT_REF_PARAM_KEY) snapshotRef: string,
-    @Query() parameters: GetScoreSnapshotRequestClass,
-    @TokenInfo() tokenInfo: AuthenticatedUser
-  ) {
-    if (snapshotRef === SnapshotsService.SCORE_COURANT_SNAPSHOT_REF) {
-      return this.referentielsScoringService.getOrCreateCurrentScore(
-        collectiviteId,
-        referentielId,
-        parameters.forceRecalculScoreCourant
-      );
-    } else {
-      return this.referentielsScoringSnapshotsService.get(
-        collectiviteId,
-        referentielId,
-        snapshotRef
-      ) as Promise<GetReferentielScoresResponseType>;
-    }
-  }
+  // TODO: À priori non utilisé. À supprimer après confirmation en prod.
+  //
+  // @AllowAnonymousAccess()
+  // @Get(
+  //   `collectivites/:${COLLECTIVITE_ID_PARAM_KEY}/referentiels/:${REFERENTIEL_ID_PARAM_KEY}/score-snapshots/:${SNAPSHOT_REF_PARAM_KEY}`
+  // )
+  // async getReferentielScoreSnapshot(
+  //   @Param(COLLECTIVITE_ID_PARAM_KEY) collectiviteId: number,
+  //   @Param(REFERENTIEL_ID_PARAM_KEY) referentielId: ReferentielId,
+  //   @Param(SNAPSHOT_REF_PARAM_KEY) snapshotRef: string,
+  //   @Query() parameters: GetScoreSnapshotRequestClass,
+  //   @TokenInfo() tokenInfo: AuthenticatedUser
+  // ) {
+  //   if (snapshotRef === SnapshotsService.SCORE_COURANT_SNAPSHOT_REF) {
+  //     return this.referentielsScoringService.getOrCreateCurrentScore(
+  //       collectiviteId,
+  //       referentielId,
+  //       parameters.forceRecalculScoreCourant
+  //     );
+  //   } else {
+  //     return this.referentielsScoringSnapshotsService.get(
+  //       collectiviteId,
+  //       referentielId,
+  //       snapshotRef
+  //     ) as Promise<ScoresPayload>;
+  //   }
+  // }
 
-  @AllowAnonymousAccess()
-  @Get(
-    `collectivites/:${COLLECTIVITE_ID_PARAM_KEY}/referentiels/:${REFERENTIEL_ID_PARAM_KEY}/score-snapshots/:${SNAPSHOT_REF_PARAM_KEY}/export`
-  )
-  async exportReferentielScoreSnapshot(
-    @Param(COLLECTIVITE_ID_PARAM_KEY) collectiviteId: number,
-    @Param(REFERENTIEL_ID_PARAM_KEY) referentielId: ReferentielId,
-    @Param(SNAPSHOT_REF_PARAM_KEY) snapshotRef: string,
-    @Query() parameters: GetScoreSnapshotRequestClass,
-    @TokenInfo() tokenInfo: AuthenticatedUser,
-    @Res() res: Response,
-    @Next() next: NextFunction
-  ) {
-    this.logger.log(
-      `Export du score du referentiel ${referentielId} pour la collectivite ${collectiviteId} et le snapshot ${snapshotRef}`
-    );
-    this.exportReferentielScoreService.sendExportReferentielScore(
-      collectiviteId,
-      referentielId,
-      snapshotRef,
-      parameters,
-      res,
-      next
-    );
-  }
-
-  @Delete(
-    'collectivites/:collectivite_id/referentiels/:referentiel_id/score-snapshots/:snapshot_ref'
-  )
-  async deleteReferentielScoreSnapshot(
-    @Param('collectivite_id') collectiviteId: number,
-    @Param('referentiel_id') referentielId: ReferentielId,
-    @Param('snapshot_ref') snapshotRef: string,
-    @TokenInfo() tokenInfo: AuthUser
-  ): Promise<void> {
-    return this.referentielsScoringSnapshotsService.delete(
-      collectiviteId,
-      referentielId,
-      snapshotRef,
-      tokenInfo
-    );
-  }
+  // TODO: À priori non utilisé. À supprimer après confirmation en prod.
+  //
+  // @Delete(
+  //   'collectivites/:collectivite_id/referentiels/:referentiel_id/score-snapshots/:snapshot_ref'
+  // )
+  // async deleteReferentielScoreSnapshot(
+  //   @Param('collectivite_id') collectiviteId: number,
+  //   @Param('referentiel_id') referentielId: ReferentielId,
+  //   @Param('snapshot_ref') snapshotRef: string,
+  //   @TokenInfo() tokenInfo: AuthUser
+  // ): Promise<void> {
+  //   return this.referentielsScoringSnapshotsService.delete(
+  //     collectiviteId,
+  //     referentielId,
+  //     snapshotRef,
+  //     tokenInfo
+  //   );
+  // }
 
   @AllowAnonymousAccess()
   @ApiExcludeEndpoint() // Not in documentation
@@ -260,7 +233,7 @@ export class ReferentielsScoringController {
     @Req() req: Request
   ): Promise<GetMultipleCheckScoresResponseType> {
     const checkHostUrl = `${req.protocol}://${req.get('Host')}`;
-    return await this.referentielsScoringService.checkScoreForLastModifiedCollectivite(
+    return await this.scoresService.checkScoreForLastModifiedCollectivite(
       parameters,
       checkHostUrl
     );
@@ -271,6 +244,6 @@ export class ReferentielsScoringController {
   @Get('referentiels/all/save-last-scores')
   async saveLastReferentielsScoreNewTable() {
     // TODO: endpoint to be removed, only used during migration
-    return await this.referentielsScoringService.saveLastReferentielsScoreToNewTable();
+    return await this.scoresService.saveLastReferentielsScoreToNewTable();
   }
 }
