--- backend/src/referentiels/models/get-score-snapshot.request.ts
@@ -1,16 +1,0 @@
-import { z } from 'zod';
-
-export const getScoreSnapshotRequestSchema = z
-  .object({
-    forceRecalculScoreCourant: z
-      .union([
-        z.enum(['true', 'false']).transform((value) => value === 'true'),
-        z.boolean(),
-      ])
-      .optional()
-      .describe(`Force le recalcul du score courant`),
-  })
-  .describe('Paramètres de la requête pour obtenir un snapshots de score');
-export type GetScoreSnapshotRequestType = z.infer<
-  typeof getScoreSnapshotRequestSchema
->;
