--- backend/src/referentiels/models/get-score-snapshots.response.ts
@@ -1,67 +1,0 @@
-import z from 'zod';
-import { SnapshotJalonEnum, snapshotSchema } from '../snapshots/snapshot.table';
-import { referentielIdEnumSchema } from './referentiel-id.enum';
-
-export const scoreSnapshotInfo = snapshotSchema
-  .pick({
-    ref: true,
-    nom: true,
-    date: true,
-    typeJalon: true,
-    pointFait: true,
-    pointProgramme: true,
-    pointPasFait: true,
-    pointPotentiel: true,
-    referentielVersion: true,
-    auditId: true,
-    createdAt: true,
-    createdBy: true,
-    modifiedAt: true,
-    modifiedBy: true,
-  })
-  .extend({
-    pointNonRenseigne: z.number().optional().describe('Points non renseignés'),
-  });
-
-export type ScoreSnapshotInfoType = z.infer<typeof scoreSnapshotInfo>;
-
-export const scoreSnapshotCollectiviteInfo = snapshotSchema
-  .pick({
-    collectiviteId: true,
-    referentielId: true,
-    ref: true,
-    nom: true,
-    date: true,
-    typeJalon: true,
-    pointFait: true,
-    pointProgramme: true,
-    pointPasFait: true,
-    pointPotentiel: true,
-    referentielVersion: true,
-    auditId: true,
-    createdAt: true,
-    createdBy: true,
-    modifiedAt: true,
-    modifiedBy: true,
-  })
-  .extend({
-    pointNonRenseigne: z.number().optional(),
-  });
-
-export type ScoreSnapshotCollectiviteInfoType = z.infer<
-  typeof scoreSnapshotCollectiviteInfo
->;
-
-export const getScoreSnapshotsResponseSchema = z
-  .object({
-    collectiviteId: z.number(),
-    referentielId: referentielIdEnumSchema,
-    typesJalon: z.nativeEnum(SnapshotJalonEnum).array(),
-    snapshots: scoreSnapshotInfo.array(),
-  })
-  .describe(
-    'Liste des snapshots de score pour une collectivité et un référentiel'
-  );
-export type GetScoreSnapshotsResponseType = z.infer<
-  typeof getScoreSnapshotsResponseSchema
->;
