+++ backend/src/referentiels/referentiels.module.ts
@@ -2,15 +2,16 @@ import { Module } from '@nestjs/common';
 import { CollectivitesModule } from '../collectivites/collectivites.module';
 import { PersonnalisationsModule } from '../personnalisations/personnalisations.module';
 import { SheetModule } from '../utils/google-sheets/sheet.module';
-import { ComputeScoreRouter } from './compute-score/compute-score.router';
 import { ReferentielsScoringController } from './compute-score/scores.controller';
 import ScoresService from './compute-score/scores.service';
-import ExportScoreService from './export-score/export-score.service';
+import { ExportScoreController } from './export-score/export-score.controller';
+import { ExportScoreService } from './export-score/export-score.service';
 import { GetReferentielController } from './get-referentiel/get-referentiel.controller';
 import { GetReferentielService } from './get-referentiel/get-referentiel.service';
 import { ImportReferentielController } from './import-referentiel/import-referentiel.controller';
 import ImportReferentielService from './import-referentiel/import-referentiel.service';
 import { GetLabellisationRouter } from './labellisations/get-labellisation.router';
+import { GetLabellisationService } from './labellisations/get-labellisation.service';
 import { LabellisationService } from './labellisations/labellisation.service';
 import { StartAuditRouter } from './labellisations/start-audit/start-audit.router';
 import { StartAuditService } from './labellisations/start-audit/start-audit.service';
@@ -19,17 +20,18 @@ import { ValidateAuditService } from './labellisations/validate-audit/validate-a
 import { ListActionDefinitionsService } from './list-action-definitions/list-action-definitions.service';
 import { ListActionsRouter } from './list-actions/list-actions.router';
 import { ReferentielsRouter } from './referentiels.router';
+import { ListSnapshotsService } from './snapshots/list-snapshots/list-snapshots.service';
 import { SnapshotsRouter } from './snapshots/snapshots.router';
 import { SnapshotsService } from './snapshots/snapshots.service';
 import { UpdateActionStatutRouter } from './update-action-statut/update-action-statut.router';
 import { UpdateActionStatutService } from './update-action-statut/update-action-statut.service';
-import { ListSnapshotsService } from './snapshots/list-snapshots/list-snapshots.service';
 
 @Module({
   imports: [CollectivitesModule, SheetModule, PersonnalisationsModule],
   providers: [
     GetReferentielService,
     ImportReferentielService,
+    ReferentielsRouter,
 
     ListActionDefinitionsService,
     ListActionsRouter,
@@ -37,29 +39,28 @@ import { ListSnapshotsService } from './snapshots/list-snapshots/list-snapshots.
     UpdateActionStatutService,
     UpdateActionStatutRouter,
 
-    // Labellisation
-    LabellisationService,
-    GetLabellisationRouter,
-    StartAuditService,
-    StartAuditRouter,
-    ValidateAuditService,
-    ValidateAuditRouter,
+    ScoresService,
+    ExportScoreService,
 
     SnapshotsService,
     SnapshotsRouter,
     ListSnapshotsService,
 
-    ScoresService,
-    ExportScoreService,
-    ComputeScoreRouter,
-
-    ReferentielsRouter,
+    // Labellisation
+    LabellisationService,
+    GetLabellisationService,
+    GetLabellisationRouter,
+    StartAuditService,
+    StartAuditRouter,
+    ValidateAuditService,
+    ValidateAuditRouter,
   ],
   exports: [ReferentielsRouter],
   controllers: [
     GetReferentielController,
     ImportReferentielController,
     ReferentielsScoringController,
+    ExportScoreController,
   ],
 })
 export class ReferentielsModule {}
