+++ backend/src/referentiels/snapshots/index-domain.ts
@@ -1,0 +1,5 @@
+// `index-domain.ts` file must only contain exports that are shareable with client-side apps.
+// Exports from this file can be imported with `@/domain/*` alias path.
+
+export * from './snapshot-jalon.enum';
+export * from './snapshot.table';
