+++ backend/src/referentiels/snapshots/scores-payload.dto.ts
@@ -3,55 +3,59 @@ import {
   CollectiviteAvecType,
   collectiviteAvecTypeSchema,
 } from '../../collectivites/identite-collectivite.dto';
-import { ActionDefinition, actionDefinitionSchema } from '../index-domain';
+import { ComputeScoreMode } from '../compute-score/compute-score-mode.enum';
+import {
+  ScoreFinalFields,
+  scoreFinalFieldsSchema,
+} from '../compute-score/score.dto';
 import {
   ActionDefinitionEssential,
   actionDefinitionEssentialSchema,
   TreeNode,
   treeNodeSchema,
 } from '../models/action-definition.dto';
-import { ComputeScoreMode } from '../models/compute-scores-mode.enum';
+import {
+  ActionDefinition,
+  actionDefinitionSchema,
+} from '../models/action-definition.table';
 import {
   ReferentielId,
   referentielIdEnumSchema,
 } from '../models/referentiel-id.enum';
-import {
-  SnapshotJalon,
-  snapshotJalonEnumSchema,
-} from '../snapshots/snapshot.table';
-import { ScoreFinalFields, scoreFinalFieldsSchema } from './score.dto';
+import { SnapshotJalon, snapshotJalonEnumSchema } from './snapshot-jalon.enum';
 
-export const getReferentielScoresResponseSnapshotInfoSchema = z.object({
-  ref: z.string(),
-  nom: z.string(),
-  createdAt: z.string().datetime(),
-  createdBy: z.string().nullable(),
-  modifiedAt: z.string().datetime(),
-  modifiedBy: z.string().nullable(),
-});
+// export const getReferentielScoresResponseSnapshotInfoSchema = z.object({
+//   ref: z.string(),
+//   nom: z.string(),
+//   createdAt: z.string().datetime(),
+//   createdBy: z.string().nullable(),
+//   modifiedAt: z.string().datetime(),
+//   modifiedBy: z.string().nullable(),
+// });
 
-export const getReferentielScoresResponseSchema = z
+export const scoresPayloadSchema = z
   .object({
     collectiviteId: z.number(),
     referentielId: referentielIdEnumSchema,
     referentielVersion: z.string(),
-    collectiviteInfo: collectiviteAvecTypeSchema,
+    auditId: z.number().optional(),
     date: z.string().datetime(),
+    jalon: snapshotJalonEnumSchema,
+
+    collectiviteInfo: collectiviteAvecTypeSchema,
     scores: treeNodeSchema(
       actionDefinitionSchema
         .pick({ identifiant: true, nom: true, categorie: true })
         .merge(actionDefinitionEssentialSchema)
         .merge(scoreFinalFieldsSchema)
     ),
-    jalon: snapshotJalonEnumSchema,
-    auditId: z.number().optional(),
     anneeAudit: z.number().optional(),
-    snapshot: getReferentielScoresResponseSnapshotInfoSchema.optional(),
+    // snapshot: getReferentielScoresResponseSnapshotInfoSchema.optional(),
     mode: z.nativeEnum(ComputeScoreMode),
   })
   .describe('Score de la collectivité pour un référentiel et la date donnée');
 
-export type GetReferentielScoresResponseType = {
+export type ScoresPayload = {
   collectiviteId: number;
   referentielId: ReferentielId;
   referentielVersion: string;
@@ -65,13 +69,13 @@ export type GetReferentielScoresResponseType = {
   jalon: SnapshotJalon;
   auditId?: number;
   anneeAudit?: number;
-  snapshot?: {
-    ref: string;
-    nom: string;
-    createdAt: string;
-    createdBy: string | null;
-    modifiedAt: string;
-    modifiedBy: string | null;
-  };
+  // snapshot?: {
+  //   ref: string;
+  //   nom: string;
+  //   createdAt: string;
+  //   createdBy: string | null;
+  //   modifiedAt: string;
+  //   modifiedBy: string | null;
+  // };
   mode: ComputeScoreMode;
 };
