+++ backend/src/utils/index-domain.ts
@@ -8,3 +8,4 @@ export * from './enum.utils';
 export * from './nest/errors.utils';
 export * from './number.utils';
 export * from './pagination.schema';
+export * from './string.utils';
