+++ backend/src/utils/string.utils.ts
@@ -1,0 +1,12 @@
+import slugify from 'slugify';
+
+export function toSlug(s: string): string {
+  if (!s) {
+    return s;
+  }
+
+  return slugify(s.toLowerCase(), {
+    replacement: '-',
+    remove: /[*+~.()'"!:@]/g,
+  });
+}
